fix:resolves issues raised by maintainers
