+++ host/src/components/Sidebar/app-sidebar.tsx
@@ -3,22 +3,58 @@ import {
   Sidebar,
   SidebarContent,
   SidebarFooter,
+  SidebarGroup,
   SidebarHeader,
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
   SidebarRail,
 } from "@/components/ui/sidebar";
-import { NavMain } from "./nav-main";
-import { NavHeader } from "./nav-header";
 import { NavUser } from "./nav-user";
-import { data } from "./data";
+import { data, getSidebarList } from "./sidebarMenuItem";
+import { Link, useLocation } from "react-router-dom";
+import { HomeHashtag } from "iconsax-react";
 
 export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
+  const pathname = useLocation().pathname;
+  const items = getSidebarList(pathname);
   return (
     <Sidebar collapsible="icon" {...props}>
       <SidebarHeader>
-        <NavHeader />
+        <SidebarMenu>
+          <SidebarMenuItem>
+            <SidebarMenuButton
+              size="lg"
+              className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
+            >
+              <div className="flex aspect-square size-8 items-center justify-center rounded-lg bg-shorttiee_primary text-sidebar-primary-foreground">
+                <HomeHashtag className="size-5" />
+              </div>
+              <div className="grid flex-1 text-left text-base leading-tight">
+                <span className="truncate font-semibold">Shorttiee</span>
+              </div>
+            </SidebarMenuButton>
+          </SidebarMenuItem>
+        </SidebarMenu>
       </SidebarHeader>
       <SidebarContent>
-        <NavMain />
+        <SidebarGroup>
+          <SidebarMenu>
+            {items.map((item) => (
+              <SidebarMenuItem key={item.title}>
+                <SidebarMenuButton tooltip={item.title} asChild className="py-6 px-4">
+                  <Link
+                    to={item.url}
+                    className={`font-medium ${item.isActive ? "bg-shorttiee_primary text-white" : ""}`}
+                  >
+                    {item.icon && <item.icon />}
+                    {item.title}
+                  </Link>
+                </SidebarMenuButton>
+              </SidebarMenuItem>
+            ))}
+          </SidebarMenu>
+        </SidebarGroup>
       </SidebarContent>
       <SidebarFooter>
         <NavUser user={data.user} />
