--- host/src/components/Sidebar/nav-header.tsx
@@ -1,22 +1,0 @@
-import { SidebarMenu, SidebarMenuButton, SidebarMenuItem } from "@/components/ui/sidebar";
-import { HomeHashtag } from "iconsax-react";
-
-export function NavHeader() {
-  return (
-    <SidebarMenu>
-      <SidebarMenuItem>
-        <SidebarMenuButton
-          size="lg"
-          className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
-        >
-          <div className="flex aspect-square size-8 items-center justify-center rounded-lg bg-shorttiee_primary text-sidebar-primary-foreground">
-            <HomeHashtag className="size-5" />
-          </div>
-          <div className="grid flex-1 text-left text-base leading-tight">
-            <span className="truncate font-semibold">Shorttiee</span>
-          </div>
-        </SidebarMenuButton>
-      </SidebarMenuItem>
-    </SidebarMenu>
-  );
-}
